@@ -14,7 +14,7 @@ actual fun createLocale(
 
     builder.setLanguage(language.lowercase())
     if (script != null)
-        builder.setScript(script.capitalize())
+        builder.setScript(script.replaceFirstChar { if (it.isLowerCase()) it.titlecase(Locale.getDefault()) else it.toString() })
     if (country != null)
         builder.setRegion(country.uppercase())
     if (variant != null)
