@@ -1,6 +1,6 @@
 package de.comahe.i18n4k.config
 
-import de.comahe.i18n4k.Locale
+import de.comahe.i18n4k.createLocale
 import de.comahe.i18n4k.messages.formatter.MessageFormatter
 import de.comahe.i18n4k.messages.formatter.MessageFormatterDefault
 import de.comahe.i18n4k.messages.formatter.provider.DeclensionProvider
@@ -22,7 +22,7 @@ class I18n4kConfigDefault : I18n4kConfig {
 
     /** atomic ref for [defaultLocale] */
     private val defaultLocaleRef: AtomicRef<Locale> =
-        atomic(Locale("en"))
+        atomic(createLocale("en"))
 
     /** atomic ref for [locale] */
     private val localeRef: AtomicRef<Locale> =
@@ -47,7 +47,7 @@ class I18n4kConfigDefault : I18n4kConfig {
 
     /** Restores the default settings. */
     fun restoreDefaultSettings() {
-        defaultLocale = Locale("en")
+        defaultLocale = createLocale("en")
         locale = systemLocale
         messageFormatter = MessageFormatterDefault
         treadBlankStringAsNull = true
